refactor: moves `middlewares` folder to `/src/application/`
