+++ src/factories/makeAuthenticationMiddleware.ts
@@ -1,4 +1,4 @@
-import { AuthenticationMiddleware } from "../middlewares/AuthenticationMiddleware";
+import { AuthenticationMiddleware } from "../application/middlewares/AuthenticationMiddleware";
 
 export function makeAuthenticationMiddleware() {
   return new AuthenticationMiddleware();
